pyproject.toml
[tool.poetry]
name = "gif2tgsticker"
version = "0.0.2"
description = ""
authors = ["ReallyiAmGhost <really@iamghost.kr>"]

[tool.poetry.dependencies]
python = "~3.9"
tkinterdnd2 = "^0.3.0"
ffmpeg-python = "^0.2.0"

[tool.poetry.dev-dependencies]
pyinstaller = "^4.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
